Implements simple verifications on the controllers
Some of this verifications block certain endpoints that in a future
could be accessible by administrators. As that's not included on the
minimum requirements, there are blocked for now with a 403.

It's not very DRY, so in a future a lot of this validations could be
made on endpoint middlewares.

+++ src/controllers/GameMatchController.ts
@@ -1,6 +1,10 @@
 import GameMatch from "models/GameMatch";
+import User from "models/User";
+import Game from "models/Game";
 
 const GameMatchModel: any = new GameMatch().getModelForClass(GameMatch);
+const GameModel: any = new Game().getModelForClass(Game);
+const UserModel: any = new User().getModelForClass(User);
 
 export default class GameMatchController {
     public static async listAll(req, res) {
@@ -14,10 +18,25 @@ export default class GameMatchController {
 
     public static async create(req, res) {
         try {
-            let toRet = await new GameMatchModel(req.body).save();
-            res.send(toRet);
+            req.body.game = await GameModel.findById(req.body.game._id);
+            req.body.host = await UserModel.findById(req.body.host._id);
         } catch (err) {
             res.status(400).send(err);
+            return;
+        }
+
+        if (req.body.host._id.toString() != req.user._id.toString()) {
+            res.status(403).send({ message: "You can only create matches in which the host is you" })
+            return;
+        }
+
+        try {
+            let toRet = await new GameMatchModel(req.body).save();
+            res.send(toRet);
+            return;
+        } catch (err) {
+            res.status(400).send(err);
+            return;
         }
     }
 
@@ -32,20 +51,43 @@ export default class GameMatchController {
 
     public static async update(req, res) {
         try {
-            let toRet = await GameMatchModel.findOneAndUpdate({ _id: req.params.gameMatchId }, req.body, { new: true });
-            res.send(toRet);
+            req.body.host = await UserModel.findById(req.body.host._id);
+            req.body.game = await GameModel.findById(req.body.game._id);
         } catch (err) {
             res.status(400).send(err);
+            return;
+        }
+
+        if (req.body.host._id.toString() != req.user._id.toString()) {
+            res.status(403).send({ message: "You can only edit a match in which the host is you" })
+            return;
+        }
+
+        try {
+            let toRet = await GameMatchModel.findOneAndUpdate({ _id: req.params.gameMatchId }, req.body, { new: true });
+            res.send(toRet);
+            return;
+        } catch (err) {
+            res.status(400).send(err);
+            return;
         }
     }
 
     public static async delete(req, res) {
+        let targetMatch = await GameMatchModel.findById(req.params.gameMatchId);
         try {
-            let toRet = await GameMatchModel.findByIdAndRemove(req.params.gameMatchId);
-            if (!toRet) {
+            if (!targetMatch) {
                 res.status(404).send({ message: 'GameMatch not found' });
+                return;
             } else {
-                res.send({ message: 'GameMatch successfuly deleted' });
+                if (targetMatch.host._id.toString() != req.user._id.toString()) {
+                    res.status(403).send({ message: "You can only delete a match in which the host is you" });
+                    return;
+                } else {
+                    await GameMatchModel.findByIdAndRemove(req.params.gameMatchId);
+                    res.send({ message: 'GameMatch successfuly deleted' });
+                    return;
+                }
             }
         } catch (err) {
             res.status(400).send(err);
